Awesome work

+++ app/Library/FileHandler.php
@@ -3,20 +3,21 @@
 namespace App\Library;
 
 use Exception;
-use Illuminate\Http\File;
-use Illuminate\Support\Facades\Storage;
+use App\Library\Str;
+use Illuminate\Support\Facades\Date;
 
 class FileHandler {
 
     static function upload($file, $path, $name){
         $ext = $file->getClientOriginalExtension();
-        $imageName = Str::slug($name).'.'.$ext;
+        $randStr = Date::now();
+        $imageName = Str::slug("$name-$randStr").'.'.$ext;
         $file->move(public_path($path), $imageName);
         return "$path/$imageName";
     }
 
     static function update($file, $path, $name, $oldFilePath){
-        self::deleteFile($oldFilePath);
+        $oldFilePath && self::deleteFile($oldFilePath);
         return  self::upload($file, $path, $name);
     }
 
@@ -26,7 +27,8 @@ class FileHandler {
         }
     }
 
-    static function deleteFile(string $file){
+    static function deleteFile(string | null $file){
+        if(!$file) return;
         $path = public_path($file);
         if(file_exists($path)) unlink($path);
     }
